requirements.txt
torch
argparse
torchvision
numpy
Pillow
tqdm
sklearn
torchsummary
torchviz
matplotlib
pandas